Refactor Auth implementation
Before this commit, the authentication was fully managed by a service
in api/auth. Not only was this improperly organized, but also hard to
maintain, as any page that requires any sort of authentication required
a lot of code in order to function; and wouldn't receive updates
whenever the account data changed (login, logout, role change)

This commit refactors that and makes api/auth only manage the
communication with the API, as it should. auth/ contains the
implementation of all the authentication-related services; and adds
hooks and expectations, which are callbacks triggered whenever the
account data changes, and compared with expectations to fire a "match"
or "mismatch" callback, whenever that is required.

This makes it fairly easy to implement a page that requires, lets say,
admin privileges.